+++ melt_the_code_dart/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.1
+
+- Renamed DartUseCaseFailure to UseCaseFailure.  Realized I could export it in the melt_the_code_flutter module.
+- Added StringExtension to support converting to different data types to support dynamic better.
+
 ## 0.2.0
 
 - Implemented the useRuntimeQuery() use case function.